requirements.txt
pytorch
torchvision
cudatoolkit=10.1